Fix(FarmRequest): Correxion de llave foranea en la validacion de reglas

+++ app/Http/Controllers/FarmController.php
@@ -14,7 +14,9 @@ class FarmController extends Controller
      */
     public function index()
     {
-        return response()->json(Farm::all());
+        return response()->json(
+            Farm::where('user_id', auth()->id())->get()
+        );
     }
 
     /**
@@ -30,7 +32,10 @@ class FarmController extends Controller
      */
     public function store(StoreFarmRequest $request)
     {
-        $farm = Farm::create($request->validated());
+        $farm = Farm::create([
+            ...$request->validated(),
+            'user_id' => auth()->id()
+        ]);
 
         return response()->json([
             'message' => 'Granja creada correctamente',
@@ -43,7 +48,18 @@ class FarmController extends Controller
      */
     public function show(Farm $farm)
     {
-        return response()->json($farm);
+        if ($farm->user_id !== auth()->id()) {
+        return response()->json([
+                'message' => 'No tienes permiso para acceder a esta granja.'
+            ],403);
+        }
+
+        return response()->json(
+            $farm->load([
+            'animals:id,name,species,farm_id',
+            'products:id,name,current_stock,farm_id',
+            'recipes:id,name,farm_id'
+        ]));
     }
 
     /**
@@ -59,6 +75,14 @@ class FarmController extends Controller
      */
     public function update(UpdateFarmRequest $request, Farm $farm)
     {
+        if ($farm->user_id !== auth()->id()) 
+        {
+            return response()->json([
+                'message' => 'No tienes permiso para modificar esta granja.'
+            ],403);
+        }
+
+
         $farm->update($request->validated());
 
         return response()->json([
@@ -72,6 +96,13 @@ class FarmController extends Controller
      */
     public function destroy(Farm $farm)
     {
+        if ($farm->user_id !== auth()->id()) {
+            return response()->json([
+                'message' => 'No tienes permiso para eliminar esta granja.'
+            ],403);
+        }
+
+
         $farm->delete();
 
         return response()->json([
